refactor: confeti 파일명 -> performance로 통일
