+++ apps/client/src/App.tsx
@@ -1,10 +1,11 @@
-import { ThemeProvider, ToastContainer } from '@confeti/design-system';
-import Router from '@shared/router/router';
 import { BrowserRouter } from 'react-router-dom';
 import { QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
-import { queryClient } from './shared/utils/query-client';
+import { ThemeProvider, ToastContainer } from '@confeti/design-system';
 import { rootStyle } from '@confeti/design-system/styles';
+
+import Router from '@shared/router/router';
+import { queryClient } from './shared/utils/query-client';
 
 function App() {
   return (
