+++ apps/client/src/pages/my/components/artist/no-artist-section.tsx
@@ -1,7 +1,7 @@
 import { useNavigate } from 'react-router-dom';
 import { Button } from '@confeti/design-system';
-import * as styles from './no-artist-section.css';
 import { routePath } from '@shared/constants/path';
+import * as styles from './no-artist-section.css';
 
 const NoArtistSection = () => {
   const navigate = useNavigate();
