+++ apps/client/src/pages/my/components/profile/user-info.css.ts
@@ -1,5 +1,5 @@
-import { themeVars } from '@confeti/design-system/styles';
 import { style } from '@vanilla-extract/css';
+import { themeVars } from '@confeti/design-system/styles';
 
 export const wrapper = style({
   display: 'flex',
