+++ apps/client/src/pages/my/page/artist/artist-more.tsx
@@ -1,7 +1,7 @@
 import { ArtistCard, Header } from '@confeti/design-system';
 import { useMyArtist } from '@pages/my/hooks/use-my-favorites';
-import * as styles from './artist-more.css';
 import { ARTISTS_DATA } from '@shared/mocks/artists-data';
+import * as styles from './artist-more.css';
 
 const ArtistMore = () => {
   const { data } = useMyArtist();
