+++ apps/client/src/pages/performance/components/artist/artist-card.css.ts
