+++ apps/client/src/pages/performance/components/artist/artist-grid.tsx
