+++ apps/client/src/pages/performance/components/artist/artist-section.css.ts
