+++ apps/client/src/pages/performance/components/artist/artist-title.css.ts
