+++ apps/client/src/pages/performance/components/expanded/expanded-section.css.ts
