+++ apps/client/src/pages/performance/components/expanded/expanded-section.tsx
@@ -1,7 +1,7 @@
 import { useState, useRef, useEffect } from 'react';
 import { ConcertArtist } from '../../types/concert';
-import MoreButton from '@pages/confeti/components/button/more-button';
-import ArtistGrid from '@pages/confeti/components/artist/artist-grid';
+import MoreButton from '@pages/performance/components/button/more-button';
+import ArtistGrid from '@pages/performance/components/artist/artist-grid';
 import * as styles from './expanded-section.css';
 
 export const MAX_VISIBLE_ARTISTS = 4;
