+++ apps/client/src/pages/performance/components/performance/performance-detail.css.ts
