+++ apps/client/src/pages/performance/components/poster/poster.css.ts
