+++ apps/client/src/pages/performance/constant/confeti.ts
