+++ apps/client/src/pages/performance/page/festival-detail.tsx
@@ -1,14 +1,15 @@
 import { useState } from 'react';
 import { useParams } from 'react-router-dom';
-import Poster from '@pages/confeti/components/poster/poster';
-import Summary from '@pages/confeti/components/summary/summary';
-import Info from '@pages/confeti/components/info/info';
-import MoreButton from '@pages/confeti/components/button/more-button';
-import PerformanceDetail from '@pages/confeti/components/performance/performance-detail';
-import ArtistTitle from '@pages/confeti/components/artist/artist-title';
-import ArtistSection from '@pages/confeti/components/artist/artist-section';
 import { FloatingButton, Footer, Spacing } from '@confeti/design-system';
-import { useFestivalDetail } from '@pages/confeti/hooks/use-festival-detail';
+
+import Poster from '@pages/performance/components/poster/poster';
+import Summary from '@pages/performance/components/summary/summary';
+import Info from '@pages/performance/components/info/info';
+import MoreButton from '@pages/performance/components/button/more-button';
+import PerformanceDetail from '@pages/performance/components/performance/performance-detail';
+import ArtistTitle from '@pages/performance/components/artist/artist-title';
+import ArtistSection from '@pages/performance/components/artist/artist-section';
+import { useFestivalDetail } from '@pages/performance/hooks/use-festival-detail';
 
 const FestivalDetailPage = () => {
   const { typeId } = useParams<{ typeId: string }>();
