+++ apps/client/src/pages/performance/types/artist.ts
